+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.19.0
+
 - Support `claude-opus-4-1` model.
 - Support `gpt-5`, `gpt-5-mini`, `gpt-5-nano` models.